+++ include/Entity/Entity.h
@@ -12,14 +12,14 @@ class Entity
 {
 public:
     // update individual stats of character
-    int updateHealth(int health);
-    void updateArmour(int armour);
-    void updatePower(int Power);
+    float updateHealth(float health);
+    void updateArmour(float armour);
+    void updatePower(float Power);
 
     // return individual stats of character
-    int getHealth();
-    int getArmour();
-    int getPower();
+    float getHealth();
+    float getArmour();
+    float getPower();
 
     // set player's character, with 'sprite' and 'character' integer as parameters
     void setCharacter(DynamicSprite* sprite, int character);
@@ -38,21 +38,21 @@ public:
     int returnChar();
 
     // return original power
-    int getOGPower();
+    float getOGPower();
 
 protected:
     // pointer to a sprite
     DynamicSprite* sprite;
 
     // stats of character
-    int health;
-    int armour;
-    int power;
+    float health;
+    float armour;
+    float power;
 
     // character integer used to identify which sprite
     int character;
 
-    int originalPower;
+    float originalPower;
 };
 
 #endif
